+++ src/console/console_get_width.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_get_width() { return system_call0(CONSOLE_GET_WIDTH_SYSCALL); }
+isize console_get_width() { return system_call0(CONSOLE_GET_WIDTH_SYSCALL); }
