+++ src/console/console_set_background.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_set_background(uint64_t color) { return system_call1(CONSOLE_SET_BACKGROUND_SYSCALL, color); }
+isize console_set_background(usize color) { return system_call1(CONSOLE_SET_BACKGROUND_SYSCALL, color); }
