+++ src/console/console_write.c
@@ -2,6 +2,6 @@
 
 #include <los/syscall.h>
 
-int64_t console_write(const char* string, uint64_t string_length) {
-    return system_call2(CONSOLE_WRITE_SYSCALL, (uint64_t)string, string_length);
+isize console_write(const char* string, usize string_length) {
+    return system_call2(CONSOLE_WRITE_SYSCALL, (usize)string, string_length);
 }
