+++ src/console/console_write_ch.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t console_write_ch(char character) { return system_call1(CONSOLE_WRITE_CH_SYSCALL, character); }
+isize console_write_ch(char character) { return system_call1(CONSOLE_WRITE_CH_SYSCALL, character); }
