+++ src/filesystem/create_directory.c
@@ -2,6 +2,6 @@
 
 #include <los/syscall.h>
 
-int64_t create_directory(const char *path) {
-    return system_call1(CREATE_DIRECTORY_SYSCALL, (uint64_t)path);
+isize create_directory(const char* path) {
+    return system_call1(CREATE_DIRECTORY_SYSCALL, (usize)path);
 }
