+++ src/filesystem/open_file.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t open_file(const char* filepath, uint64_t flags) { return system_call2(OPEN_FILE_SYSCALL, (uint64_t)filepath, flags); }
+isize open_file(const char* filepath, usize flags) { return system_call2(OPEN_FILE_SYSCALL, (usize)filepath, flags); }
