+++ src/filesystem/remove_directory.c
@@ -2,6 +2,6 @@
 
 #include <los/syscall.h>
 
-int64_t remove_directory(const char* path) {
-    return system_call1(REMOVE_DIRECTORY_SYSCALL, (uint64_t)path);
+isize remove_directory(const char* path) {
+    return system_call1(REMOVE_DIRECTORY_SYSCALL, (usize)path);
 }
