+++ src/process/execute.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-ProcessID execute(const char* filename, const char** argv, const char** envp, const StandardIO* stdio) { return system_call4(EXECUTE_SYSCALL, (uint64_t)filename, (uint64_t)argv, (uint64_t)envp, (uint64_t)stdio); }
+ProcessID execute(const char* filename, const char** argv, const char** envp, const StandardIO* stdio) { return system_call4(EXECUTE_SYSCALL, (usize)filename, (usize)argv, (usize)envp, (usize)stdio); }
