+++ src/process/kill_process.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-void kill_process(int64_t id) { system_call1(KILL_PROCESS_SYSCALL, (uint64_t)id); }
+void kill_process(isize id) { system_call1(KILL_PROCESS_SYSCALL, (usize)id); }
