+++ src/process/wait_process.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t wait_process(ProcessID id) { return system_call1(WAIT_PROCESS_SYSCALL, id); }
+isize wait_process(ProcessID id) { return system_call1(WAIT_PROCESS_SYSCALL, id); }
