+++ src/thread/wait_thread.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t wait_thread(ThreadID id) { return system_call1(WAIT_THREAD_SYSCALL, id); }
+isize wait_thread(ThreadID id) { return system_call1(WAIT_THREAD_SYSCALL, id); }
