+++ src/time/get_epoch_time.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t get_epoch_time() { return system_call0(GET_EPOCH_TIME_SYSCALL); }
+isize get_epoch_time() { return system_call0(GET_EPOCH_TIME_SYSCALL); }
