+++ src/time/get_process_time.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t get_process_time() { return system_call0(GET_PROCESS_TIME_SYSCALL); }
+isize get_process_time() { return system_call0(GET_PROCESS_TIME_SYSCALL); }
