+++ src/time/get_time_zone.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-int64_t get_time_zone() { return system_call0(GET_TIME_ZONE_SYSCALL); }
+isize get_time_zone() { return system_call0(GET_TIME_ZONE_SYSCALL); }
